+++ gimmemotifs/maelstrom/__init__.py
@@ -21,7 +21,7 @@ import numpy as np
 import pandas as pd
 import seaborn as sns
 from matplotlib.gridspec import GridSpec
-from scipy.cluster.hierarchy import dendrogram, linkage, leaves_list
+from scipy.cluster.hierarchy import dendrogram, leaves_list, linkage
 from scipy.spatial.distance import pdist
 from scipy.stats import pearsonr
 from sklearn.cluster import FeatureAgglomeration