+++ CHANGELOG.md
@@ -5,7 +5,10 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 
 ## Unreleased
 
+### Fixed
 
+- Scanner.scan now returns sequence.ids along with results.
+  - if sequences cannot be parsed, both ids and results are omitted.
 
 ## [0.18.2] - 2026-04-03