+++ CHANGELOG.md
@@ -3,6 +3,8 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 
+
+
 ## Unreleased
 
 ### Changed
@@ -17,7 +19,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 ### Removed
 
 - deprecated functions `pwm_scan` and `pwm_scan_all`
-
+- warning blacklists
 
 
 ## [0.18.2] - 2026-04-03