+++ CHANGELOG.md
@@ -7,6 +7,10 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 
 ## Unreleased
 
+### Added
+
+- `gimme motifs` & `gimme maelstrom` now accept argument `--noprogress` to silence tqdm progress bars
+
 ### Changed
 
 - minor Scanner.set_threshold optimization
@@ -20,6 +24,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 
 - deprecated functions `pwm_scan` and `pwm_scan_all`
 - warning blacklists
+- unused functions `_moap_with_bg()` and `_moap_with_table()`
+- outdated copyright comments (copyright claims still found in docs and COPYING file)
+
 
 
 ## [0.18.2] - 2026-04-03